pyproject.toml
[tool.poetry]
name = "blog-api"
version = "0.1.0"
description = ""
authors = ["Zachary Juang <zachary822@me.com>"]
readme = "README.md"
packages = [{include = "api"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.0"
uvicorn = "^0.23.2"
pydantic = "^2.3.0"
motor = {extras = ["srv"], version = "^3.1.1"}
pendulum = "^2.1.2"
pyyaml = "^6.0"
newrelic = "^9.0.0"
lxml = "^4.9.1"
markdown = "^3.4.1"
strawberry-graphql = {extras = ["fastapi"], version = "^0.205.0"}
pydantic-settings = "^2.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-asyncio = "^0.21.0"
pyinstrument = "^4.3.0"
httpx = "^0.24.0"

[tool.mypy]
plugins = ["pydantic.mypy", "strawberry.ext.mypy_plugin"]
ignore_missing_imports = true
show_error_codes = true
exclude = ["tests"]

[tool.ruff]
line-length = 120

[tool.black]
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.isort]
profile = "black"
line_length = 120
filter_files = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
